+++ projects/auth/src/lib/adaptor/auth-api.adaptor.ts
@@ -1,31 +1,7 @@
 import { Injectable } from '@angular/core';
-import { Adaptor } from '../interfaces/adaptor';
+import { Adaptor, AuthAdapted, AuthApiAdaptor } from '../interfaces/adaptor';
 
-export interface AuthAdapted {
-message:string,
-token: string,
-user:  
-{
-  _id: string,
-username:string,
- firstName: string,
-lastName: string
-createdAt: string,
-email: string,
-isVerified: boolean,
-passwordResetCode: string,
-passwordResetExpires: string,
-phone: string,
-resetCodeVerified: boolean,
-role: string,
-}
-}
 
-export interface AuthApiAdaptor {
-  message:string,
-  token: string,
-  email: string
-}
 @Injectable({
   providedIn: 'root',
 })
